Assignment 4 for database systems class
Check in includes all queries specified under the "lab 4" assignment.

Tried to use subqueries as much as possible here and was able to except
for query number 4. Ended up using an intersect within the subquery to
get customers who ordered both product p01 and p07. Not using an
intersect led to it retrieving a customer who had ordered product p01 OR
product p07 which wasn't giving the desired results. Wasn't sure of
another way to retrieve information for "and" versus "or".